Added material requirement information to add-on specs and used it in the crafting UI. To craft an add-on, the player needs to have the required amount of materials in the inventory. The required materials are removed from the inventory when the add-on is crafted.
The material requirements and bonuses are now shown in the description text. Text widget is used for showing the text since scrolling is needed by items with longer descriptions.

With these additions, the add-on crafting UI is feature complete. There's still some balancing to do but that's easy to do separately.

Patch by Ari Mustonen.

git-svn-id: https://freedroid.svn.sourceforge.net/svnroot/freedroid@3080 524c2bfb-fa3f-0410-b5a2-92f4e1dc5b61