The skill icon is being drawn over the upgrade UI in 640x480 mode.
The icon of the current skill and the health and temperature bars being drawn
over the item upgrade UI and the add-on crafting UI in 640x480 mode. To fix
the problem, the drawing order is changed so that the HUD parts are drawn before
the upgrade UI and the crafting UI.
The patch also removes some 640x480 specific hacks that became useless and moves
the HUD drawing functions not having anything to do with the banner out of
DisplayBanner().
Patch by Ari Mustonen.


git-svn-id: https://freedroid.svn.sourceforge.net/svnroot/freedroid@3148 524c2bfb-fa3f-0410-b5a2-92f4e1dc5b61