LevelEditor: to-be-dropped thing is not changed after a category swap.
If the user starts a task (create a waypath, for instance) and selects
an other object's category (floor, for instance), without ending his first
task, then on the next click on the map, a new waypath creation is started.
In this patch, I force the user to finish an operation (draw a rectangle,
a line of walls, a waypoint route...) when he changes the current category.
Patch by Samuel Pitoiset.


git-svn-id: https://freedroid.svn.sourceforge.net/svnroot/freedroid@3152 524c2bfb-fa3f-0410-b5a2-92f4e1dc5b61