lvledit: Item labels are displayed.
When one zooms out and drops an item on the ground (probably one has to press
z ingame first to make item names being displayed) the name's position is not
equal to the position of the item, the name of the item is displayed at the 
place where the item would be without zooming.
This patch removes the display of item labels, in the editor.
Patch by Samuel Pitoiset.


git-svn-id: https://freedroid.svn.sourceforge.net/svnroot/freedroid@3168 524c2bfb-fa3f-0410-b5a2-92f4e1dc5b61