Tux animation (3): Tux is not facing front in the takeover minigame.
The current rotation and attack phase of Tux are used when displaying Tux
in the takeover minigame. So, depending on its current state, sometimes Tux
is facing back.
iso_put_tux() has been reworked a bit, in order to use appropriate default
values to render Tux, when animation is not wanted.


git-svn-id: https://freedroid.svn.sourceforge.net/svnroot/freedroid@3175 524c2bfb-fa3f-0410-b5a2-92f4e1dc5b61