dialogs: dixon to tell tux that he can get bot parts at blue brother if not yet met and craft addon screen clicked
This patch makes it possible for Dixon to tell Tux where he can get information
on bot parts if the craft-add-on dialog part is clicked and Tux did not yet say
the "I'd like to know how to extract bot parts" part while talking to Engel.
Patch by Matthias Kruger.


git-svn-id: https://freedroid.svn.sourceforge.net/svnroot/freedroid@3262 524c2bfb-fa3f-0410-b5a2-92f4e1dc5b61