Actually identify the killer of your friend.
Change the killertype from using the "type" to the "id" (which is fairly minor,
since mostly this only cares if it was Tux).
Since we have the id, and there is a function to look up the bot, we can then
display the bot name.
Also, makes it so Tux is not blamed for autogun deaths.
Patch by Miles.

git-svn-id: https://freedroid.svn.sourceforge.net/svnroot/freedroid@3288 524c2bfb-fa3f-0410-b5a2-92f4e1dc5b61