Define Tux's motion classes and how to order the rendering of Tux's part in a Lua config file.
Currently, we have 2 hard-coded sets of rendering orders, one for gun animation
and one for sword animation.
In order to ease the addition and configuration new animations, specification
of Tux's motion classes and part orderings has been externalized in the
tuxrender_specs.lua file.
freedroid.item_archetypes is also extended, so that one can define the motion
class to use for each weapon.

git-svn-id: https://freedroid.svn.sourceforge.net/svnroot/freedroid@3364 524c2bfb-fa3f-0410-b5a2-92f4e1dc5b61