Fix "Guard dies if pulled into shooting range of Dixons autogun".
When the game starts, there is a guard following Tux until talking to Spencer.
It is possible to "pull" this guard into the shooting range of the autogun in
Dixons room, the guard dies then.
Solved by only having the gun turn on after the player talked to Spencer.
Patch by Miles.


git-svn-id: https://freedroid.svn.sourceforge.net/svnroot/freedroid@3381 524c2bfb-fa3f-0410-b5a2-92f4e1dc5b61