Dialogs: Fix to "possible to re-start tanias dialog when she is in town".
If start_chat() is called in the middle of a dialog's node, strange things
can happen.
That's the case in Tania dialog, where start_chat() is used to switch to 
Pendragon dialog (and vice versa).
If start_chat() is used, it has to be the last thing called in a dialog's node.
Patch by Miles.


git-svn-id: https://freedroid.svn.sourceforge.net/svnroot/freedroid@3382 524c2bfb-fa3f-0410-b5a2-92f4e1dc5b61