De-hardcodification of wearable items.
Add a 'Tux part instance' in the item specifications.
This attribute is used as a prefix to find the animation archive file
needed to render a Tux part.
The default instance names are defined in tuxrender_specs.lua.
Those are used when no item is equipped on a given Tux part (for exemple,
no boots).

git-svn-id: https://freedroid.svn.sourceforge.net/svnroot/freedroid@3863 524c2bfb-fa3f-0410-b5a2-92f4e1dc5b61