shop: fix shop pricing display, change way items are calculated.
Fix the price for items that grouped together so that the pricing per unit, as well as the price for all units is shown.
Likewise, there was an error where low value (<3 credit) items (laser ammo), could not be sold due to a rounding error.
This changes selling so that it is always rounded down, but !=0 (so items the should be sellable, can be sold).
Fixes #3157909 and #3166018. Patch by Miles.

git-svn-id: https://freedroid.svn.sourceforge.net/svnroot/freedroid@4112 524c2bfb-fa3f-0410-b5a2-92f4e1dc5b61