lvledit: use display_image_on_screen to display the elements in the toolbar instead of relying on custom functions that shall be removed.
A small hack had to be done: display_image_* functions take the offset of the image into account, but in the case of the lvledit toolbar we are displaying isometric (non-zero offset) images in a 2D GUI (zero offset).  So the coordinates passed to display_image_on_screen have to be mangled a bit in this case.

git-svn-id: https://freedroid.svn.sourceforge.net/svnroot/freedroid@4184 524c2bfb-fa3f-0410-b5a2-92f4e1dc5b61