Initialize 'next time screen size' to the default screen size.
If the game is aborted due to a "fatal" error, the current game configuration
is saved, and the saved screen size is set to 'next_time screen size.
If for any reason, the game aborts during the very first loading steps, then
the current game configuration is the default one (defined in
ResetGameConfigToDefaultValues()).
The default 'next_time screen size' was not set, thus leading to a (0, 0)
screen size being saved.

git-svn-id: https://freedroid.svn.sourceforge.net/svnroot/freedroid@5048 524c2bfb-fa3f-0410-b5a2-92f4e1dc5b61