Load game configuration file before to initialize SDL.
The OpenGL context has to be initialized before to call any OpenGL drawing
primitive. To initialize the OpenGL context, we first need to call SDL_SetVideoMode(),
which needs to know the screen size to use. The screen size is defined in the
config file.
When loading the configuration file, an error can be detected, and ErrorMessage()
is then called to report it. When the game is run in OpenGL mode, ErrorMessage()
calls some OpenGL drawing primitives.
The OpenGL context being not yet initialized at that time, the game crashes...

In ErrorMessage(), we already check if SDL is initialized, and if not we do
not call OpenGL drawing primitives.
So, loading the configuration file *before* to initialize SDL fixes the issue.

git-svn-id: https://freedroid.svn.sourceforge.net/svnroot/freedroid@5049 524c2bfb-fa3f-0410-b5a2-92f4e1dc5b61