Improve Lua error message parsing.
When a lua script is run, the error string is analyzed in order to improve its
display on the console (colorization, line number added...).
This error string parser expects a line number to be included in the error message.

However, in some cases, there is no such line number, and the parser fails.

This patch improves the search of a line number in the error message to avoid a
possible crash (the crash depends on the memory content, due to the lack of a 
test against "end-of-string" when analyzing the error message).


git-svn-id: https://freedroid.svn.sourceforge.net/svnroot/freedroid@5077 524c2bfb-fa3f-0410-b5a2-92f4e1dc5b61