Prevent user interaction with the game when the world is frozen.
The new widget system is event-based. But user's interaction on the game map
as well as with some UI panels (those not using the new widget system) is
state-based. It is thus impossible for a 'new' UI to catch an interaction event
and prevent it to be 'propagated' to the game map.
This leads to the player being able to interact with game objects when the new
quest browser is open, for example, even if the quest browser is meant to cover
the whole screen (bug #513)...

This patch is a 'temporary' solution, calling world_frozen() where needed to
prevent un-wanted interactions.

git-svn-id: https://freedroid.svn.sourceforge.net/svnroot/freedroid@5141 524c2bfb-fa3f-0410-b5a2-92f4e1dc5b61