Chat screen widgetization: Use a stack of chat contexts.
Gather all the variables needed to run and display the chat screen into a new
struct chat_context.
Use a stack of chat contexts, to replace the 'run_chat() recursion' created by
subdialogs or start of a new dialog inside an other one.
Implement a simple FSM in run_chat() to select the lua script to run depending
on the current state of the chat context.