Chat screen widgetization: Run dialog scripts in coroutines.
Some lua functions used in dialog scripts call a C function that needs a user
interaction (for instance: lua_chat_says() calls chat_add_response() which calls
wait_for_click()). Currently, wait_for_click() contains its own loop to display
and interact with the chat screen. chat_do_menu_selection() is an other example.

In order to remove those specific loops and use one single loop to handle the
chat screen (the one in run_chat()), we have to be able to stop the lua VM, get
back to the dialog engine interaction loop, and resume the lua VM after the
user's interaction. This can be done with lua coroutines.

This patch introduces a load_lua_coroutine()/resume_lua_coroutine() pair, to
load and run a lua script in a coroutine, and the FSM of the dialog engine is
adapted to use them for the dialog scripts.