widget_text_list: Improve text hovering.
When the mouse is moved over the list of text entries, there is a kind of lag
(the hovered entry is not immediately highlighted). This is due to a bunch
of events being received (and handled) when the user move her mouse.

with this patch, when a mouse motion event is received, all mouse motion
events are removed from the SDL event queue, and the latest one is used
to get the actual current mouse position.