Chat screen widgetization: Remove all the remaining dedicated interaction code.
This patch makes a full use of the widgets' interaction mechanisms to implement
the user interaction on the chat screen.
It removes the dedicated code used to scroll the chat log and the chat selector,
as well as the code used to detect that the user has selected a dialog option.