teleport: free reverse
Before, to reverse the teleportation to town, we needed to use the Sanctuary skill (costs heat) or a Teleport Homing Beacon (costs the item).
The new implementation allows Tux to teleport back to the exit location outside of the town by walking into the teleporter located in  town.
The color of the town teleporter changes to blue if we can travel back for free.