Fix obstacle list defragmentation.
With current code, if the fragmented list is { -1, -1, A, -1, B } then
the defragmented list is { B, -1, A }.
Obstacle extensions are correctly changed to reflect the new position of their
associated obstacle in the defragmented list.
For instance, the extension of obstacle A refers to the 3rd slot of the
obstacle list.

The empty slots are removed when the obstacle list is saved. So the saved list
is { B, A }.

However, obstacle extension references are not changed (and can not easily be
changed) when the empty slots are removed.
As a consequence, the extension of obstacle A is saved with a reference to the
3rd slot, which is no more the right one (A is now the second obstacle in the
list).

This patch slightly changes the way the obstacle list is scanned during
defragmentation, to ensure that there are no empty slots in the defragmented
list.