Change the old map label position getter, ResolveMapLabelOnShip, by a new function.
map_label_get_map_center: returns the center (return struct). It uses gps struct instead of location. It returns a value instead of use pointer.

The new function is needed because the code uses the center of map label but the old function returned the first corner.
To get the center, a magic number is added (0.5) to the corner. This works because map labels have a fixed size of 1. We need to solve this problem before add big map_labels.