repair skills/ shops: Add the repair equipment book to the two book shops and update the repair skill.
If Tux's repair ability is at the full skill level make it so that most of the time the item is fully repaired. Give it a 5 % chance
of improving the item and a 3% chance of making it worse. Otherwise give him a skill level *10 percent chance of repairing the item completely.