Add a function to get the obstacle type by name.
Loop on the obstacle spec array to find the type of a obstacle spec by the name specified as parameter. The obstacle type is the index of obstacle spec, used as type in
the struct obstacle. The function returns -1 if no specs are found and displays a warning.

You don't need to add a name for each obstacle now.