Some external tools (such as transifex) dislike our use of non alphanumeric characters to change the font used to render a string (\1..\5).
This patch adds an other way to change the font, by using a bb-code
like notation:
[r] = \1
[w] = \2
[n] = \3
[s] = \4
[v] = \5

The two notations are available, and can be used together.
i18n strings have been changed to use the new notation.

Note that is not a real bb-code notation, which would need to rewrite
part of our text rendering code. So consider this patch as a simple
workaround...