[Lua bindings] It was a bad idea to set Tux (the FDtux instance) as a global Lua variable.
This works for Tux because there is only a single instance of FDtux,
but the use of a global var will not work for the instance of the FDnpc
(not yet written) associated with the dialog.
Let's call that insta nce MyNPC. If MyNPC is a global var, and if a node
script launches a dialog with an other NPC, then MyNPC will be overriden...

So instead of a global var, this patch defines Tux as a local upvalue of
each dialog.

Changes adapt_to_FDtux script accordingly.