Add a real-time ingame profiler (code name: "rtprof"). This is a dev tool, used to graph values in real-time.
To generate a graph, a 'probe' has to be created in the code.
4 kind of probes are currently defined: counter, timer, 1D statistics
and 2D statistics.

The rtprof is not built by default. Use 'configure --enable-rtprof' to
embed it in the code.
LeftAlt-p switches on/off the rtprof.
LeftAlt-o clears the probes.
(note that without adding probes in the code, nothing is displayed...)

See the Doxygen Modules section for documentation.