[LUA] Add code to use C closures instead of standard C functions as lua constructors. Those C closures have one upvalue (a lightuserdata), which can be used to share a C pointer between the fdrpg core and a lua constructor.
Typical usage: Set the upvalue to a pointer to a C structure, run
a lua file to load a config data (defined in a table), and have the
table constructor (C code) use the shared pointer to fill in the C
struct.

This would avoid the need for global data.