Add a function to find a localized version of a file (find_localized_file()).
The localized versions are to be put in subdirs, using locale names.
For instances, map/titles/fr ou map/titles/de.

As with gettext(), generalizations of the locale name are tried
in turn. So if the locale is 'fr_FR', but the 'fr_FR' subdir does
not exists, then the 'fr' subdir is checked.