Add a WaitNoEvent() function, to be used to wait for the end of pending events before to leave a UI panel.
Some UI screens (the dialog screen, for instance) have their own interaction
loop, and do not use the main input_handle() function.
When such a UI screen is closed, the state of the input devices (pressed keys,
cursor position and so on) can be incoherent with the interaction state
internally computed and stored by input_handle().
This function waits for a 'neutral' interaction state, and is to be used before
to return to the main game screen.