Enforce the recognition of numkeys on some non-US keyboards.
On some keyboards (AZERTY is an example), the shift modifier is needed to
access the numeric keys, which are at the top of the alphanumeric keyboard.
In this case, the standard 'toupper()' method used to get the actual value of
the pressed key does not work.
To circumvent the problem, we use the unicode value of the key, and transform
it into a keypad sym value.