Fix freedroidRPG to compile with clang.
In the GUI subsystem, we use anonymous-like functions.

With gcc, they are implemented with a nested function and a
compound statement returning a pointer to the nested fuction.
These are 2 gcc specific extensions.
Clang does not implement nested functions, but it has a 'block'
extension that can be used to implement anonymous functions.
However, a 'block function' is not compatible with a function pointer.

(See http://en.wikipedia.org/wiki/Anonymous_function)

This patch defines a WIDGET_ANONYMOUS macro to hide the 2
implementations.
WIDGET_EXECUTE and WIDGET_UPDATE_FLAG_ON_DATA macros are removed.

All update() callbacks are rewritten to use WIDGET_ANONYMOUS.

The doxygen doc is adapted accordingly.

Note: in order to compile with clang on Linux, the BlocksRuntime clang
module has to be installed (see http://compiler-rt.llvm.org/)