Improve targeting of enemies in APressed mode or with the "Use Weapon" skill.
Currently, the floor location corresponding to mouse position is used as
the target location.
When targeting an obstacle, for instance, the user tends to point at the
center of the obstacle's image instead of clicking on the floor near the
obstacle.
He thus often misses the obstacle.

This patch checks if there is an obstacle or an enemy beneath the mouse
cursor, and if so select it as the target.