Our bitmap fonts are subsets of the ISO-8859-15 8-bit codeset. In order to be able to use non Latin-9 alphabets, this patch adds handling of charset encodings.
The encoding to use for a given language is defined in
map/languages.lua. If, for a given language, an encoding is not defined, it
defaults to ASCII.

A new structure of the graphics/font subdir is used:
- fonts in graphics/font are 7-bit ASCII fonts.
- fonts in 8-bit encodings are placed into graphics/font/<codeset name>
- if a font file is not found in a <codeset name> subdir, the ASCII version is
  used

(This patch does not include any change to graphics/font, but this is not an
 issue due to that 3rd point)

For instance, our current fonts will be moved into graphics/font/ISO-8859-15

When a user changes the current language via the 'Languages' menu, the
current encoding used to read and convert gmo files is changed accordingly.

Note that currently the fonts are not reloaded, the game has to be
restarted. A future patch will enable in-game reloading of fonts.