Add gluem/make_bmchars, a tool to create bitmap individual chars from a TTF, conditionally built if SDL_ttf is installed. Mention the need for SDL_ttf into INSTALL.freedroidRPG.
Note: There is a bug when generating some outlined glyphs: if
there is no room on the right of the glyph, the ouline pixels are
wrapped to the left of the bitmap. That's perhaps a SDL_ttf bug, and
then we should switch to directly use libfreetype.
In the meanwhile, we will use adapted TTFs to avoid that bug.