Add a flag to denote if an image's texture is a 'sub-texture'.
For reading an atlas or a font bitmap, an OpenGL texture of the
whole image is created.
Later on, 'struct image's are created for the individual parts.
The whole texture is also used by those individual 'struct image's
(unless we were not able to create the whole texture).
With this patch, a flag is used to denote that the whole texture
is reused in individual parts.
This will be used to avoid to delete several times the same texture
when individual images are deleted.
Example use case: deleting a bitmap font to reload a new bitmap with an
other charset encoding.