Use vpath directives in tools/ Makefiles, to compile all source files locally (even ../../src/pngfuncs.c).
When building in a separate directory, a VPATH to $(srcdir) is
set. If a .o file already exists in the $(srcdir) tree, it will
satisfy the dependency, due to VPATH. But the local .o file will
be missing when linking the final exe.

So before to build in a separate directory, all .o files have to be
removed from the $(srcdir) tree.

In some circumstances (for instance if a tool was built before the
main exe), pngfuncs.o was created in the $(srcdir)/src, making the
whole build failing.

Using the vpath directive, this patch ensures that building a tool
will locally compile pngfuncs.c