Add 'volatile obstacle' concept, and functions to handle them. Blood spread ion floor when a bot is hit or killed is an example of a volatile obstacle.
Volatile obstacles are stored in a linked list associated to each map's
tiles. They have a vanish time, after which they disappear.

A new 'vanish_delay' attribute is added to obstacle's spec
(obstacle_specs.lua). It defaults to 10 seconds. A volatile
obstacle has to have a IS_VOLATILE flag.

The volatile obstacles of a level are immediately removed when the
level is respawned.

Note: an obstacle flagged as IS_VOLATILE but placed on floor by the
level editor is not considered as a 'volatile' one, and so it will
not be removed after some times or when the level is re-spawned.