[widgets] Replace use of 'lambda-like' functions by use of standard functions, and adapt configure.ac and the widget documentation accordingly.
The update callback of widgets is implemented using lambda-like
functions (through the use of the WIDGET_ANONYMOUS macro).

When gcc is used to build the game, those lambda-like functions are
implemented using nested functions inside statement expressions.
Gcc implements nested functions with 'trampolines' (code added
on the stack at run-time, which calls the actual function, see
https://gcc.gnu.org/onlinedocs/gcc/Nested-Functions.html and
https://gcc.gnu.org/onlinedocs/gccint/Trampolines.html).

The trampoline code being on the stack, it is not more available
as soon as the containing function is leaved. A pointer to the
nested function (the trampoline) is thus invalid outside of the
containing function.

However, when a nested function does not use 'upvalues', the call to
the trampoline code can be optimized to a direct call to the actual
nested function code (in the .text section).

Prior to gcc 5, that optimization was done even in -O0. That's no more
true with gcc 5 (see the small test below).

This means that we can no more use nested functions to implement
our WIDGET_ANONYMOUS. No other method being available, we have to
get back to the use of a 'standard' function pointer.

------------------ Small test ------------------

$ cat test.c
void main(int argc, char *argv[])
{
  void inner_f() { printf("tada\n"); }
  printf("%p\n", inner_f);
}

$ gcc4.9 -O0 test.c -o test
$ nm test | grep inner_f
0000000000400580 t inner_f.2180
$ ./test
0x400580

$ gcc5.1 -O0 test.c test
$ nm test | grep inner_f
0000000000400586 t inner_f.2287
$ ./test
0x7fffba0a14a0

$ gcc5.1 -O1 test.c test
$ nm test | grep inner_f
0000000000400586 t inner_f.2323
$ ./test
0x400586