lvleditor: Stop drawing a line of walls when the mouse pointer leaves the level's map.
Currently, the walls ouside the level are not drawn, when the mouse
leaves the level, but the lvleditor is still in 'drawing line of walls'
mode. As soon as the mouse re-enters the level's boundaries, the
level editor tries to plant a new wall, but on an invalid position.

This patch calls end_wall_line() when the mouse is not in the level.