Extend the endgame, work toward end of act 1.
This extends the game past the point where the player propagates the
firmware update, killing all bots near the town and disabling the Hell
Fortress.
The firmware update disables all security measures and exposes a secret
teleportation network that enables Spencer to teleport into the server
room, accompanied by Bender, who makes a point about stories and what
they mean to us, and the escort from the beginning of the game, who
offers his own minimalistic congratulations. The player can ask the
teleportation network to be linked to the town's teleporter for easy
travelling back and forth. The game now ends with Spencer about to share
his plan for the future with the player.
After returning to the Town, Iris shares an ominous revelation, Tamara
mourns her role in history, and Engel vows revenge. Also, the 614 in the
cryo facility proves it is not blind.

This patch is only a start; the true end of act 1 will involve changes
to nearly all dialogs in the game, and more patches will follow.