Fix issue829.
Pushing the firmware update now visibly causes security systems to crash, and locks the door to the server room. The player can no longer leave through it, and must check the new OS version in the terminal dialog to proceed.
This also prevents Bender from opening and closing this door when passing next to it.