Improve Freedesktop compliance:
Package names in 'standard' distributions are in lowercase, so
we change the names used in the freedesktop spec files, as well as the
icon file names, to conform to this rule.

Add a screenshot image in the pkgs/freedesktop subdir, and use the link
to the raw file available in the git repo rather than to a file hosted
on the website (easier to change it if needed).