Multi-platform specification of the windres tool's pathname.
When building a Windows version of the game, the windres tool is used
to create a Windows ressource file containing the application icon.

The actual name of that tool depends on the building environnement.

When building with msys2/mingw64 on Windows, the tool is called
'windres'.

When cross-building with mingw64 on Linux, the tool name can be
prefixed by the target-triplet (x86_64-w64-mingw32-windres, for
instance).

This patch uses the autoconf AC_PATH_TARGET_TOOL macro, whose purpose
is to find the path of a tool in such a case.