Introduction of a framework for savegame converters, and use it to fix a savegame bug in the 0.16 version..
Those converters, composed of piped filter functions, can be
used to fix bugs in savegames, or to convert from an old version
to a new one, provided that the changes are not too complex.
Do not expect this framework to be able to easily resolve our
well known 'savegame incompatibility' issue...

This framework is used, in this commit, to fix a bug in savegames
from the 0.16 version, when German language is used (Lua error
when parsing the .sav file). See savegame/filters_0_16.c for
more information about that bug.