When built with clang, the level validator reports many errors about some waypoints being too close.
This is due to the fast-math compilation flag, because fast-math implementation
of clang computes strangely: sqrt(1) = 0.9999999404

This is expected with fast-math, but this leads the level validator to report
false errors.

This patch adds an EPSILON to the waypoint distance test, to take into
account a potential 'imprecision' of sqrt() computations.