Dialog validation: fix an error on long dialog script pathnames.
We currently use the lua debug short_src data to get the filename
of a running lua script. But short_src is clamped to 60 characters.
Since the data hierarchy restructuration, some pathnames are larger
than 60 characters, and so the dialog validator fails.

The same kind of error was reported when running the validation on
an installed game (issue801).

This is fixed in this patch by using 'source' instead of 'short_src'.