This skill increase the chance to drop items by killing bots or by opening chests).
The skill is taught by Iris, as requested by Xenux in bug 548 .
Basically, this adds a new skill "treasure" which each level increases the chance
that a bot is going to drop an item by 2 percentage points.
So you can improve from 10 to 20 % the chance of a bot dropping an useful item.

Iris teaches tux up to five ranks from this skill.