Extend the dynarray concept to manage sparse dynarrays.
Some items, such as bullets, stored in an array, are frequently created
and deleted.

Removing a member of a 'standard' dynarray involves to move in memory a
part of the array, which can be a costly operation.

A sparse dynarray is an array possibly containing empty slots.
A 'used_members' array is added to 'struct dynarray', containing
booleans to define of the member of the same index is actually used or
not.

New functions are introduced to create a sparse dynarray:
- sparse_dynarray_init()
- sparse_dynarray_alloc()

All other dynarray funtions (dynarray_add(), dynarray_del(), ...) now
internally handle the two kind of dynarrays.
So only the initialization is specific: a sparse dynarray is used the
same way than a non sparse dynarray.

A third function is added, to check if a member of a sparse dynarray is
empty or not: sparse_dynarray_member_used()