Fix display of vanishing text in the level editor.
A text vanishing after some time can be displayed on the lower left of
the screen (when right-clicking on an item on the toolbar, or when
changing the zoom factor).

Some constant values are currently used to define the 'vanishing time'.
Those values are too low for very low framerates (10fps, for instance),
leading to the vanishing text being not displayed at all.

This patch uses the current framerate to compute the 'vanishing time'.

The place used to display vanishing messages is also used to display
a warning when editing a random level.
The current code does not check if a vanishing text is to be displayed,
and so it is 'overloaded' by the warning text.

This is fixed in this patch, by changing the condition used to display
the warning text.