Tux self-destruct blasts found outside map boundaries.
The blast z position seems to not be set in start_tux_death_explosions()
function. So the created blast will use the level that was already
stored in the reused array slot.

This patch also changes a bit the formula used to compute the random
position of the blasts, to avoid "strange black magic constants".