Widgetization of 'title screen'.
Add a widget_autoscroll_text widget, that display a text which
automatically scrolls from bottom to top.

Use this widget to create a widget based title_screen element
(in the new title.c file).

Change a bit the GUI to add an 'Exit' button.

Use title_screen in the code that play a title.