Add functions to play a voice acting sound file.
To play a voice acting file, we would need to mix an other SDL Music
stream, but SDL can only handle one Music stream at a time.

The other sound related functions currently available are designed
to SFX sounds and are not well fitted for voice acting, mainly
because there is no need to keep a voice acting file into a cache.
Moreover, we want to be able to fade out a voice acting if an other
one is started.

Consequently, a set of functions have been specifically designed to
play/stop a voice: play_voice(), resume_voice() and stop(voice).