GL: flush struct image batch before drawing health bars
This fixes a bug where the bar, which is normally drawn after the
obstacles it's supposed to appear on top of, will be drawn before, due
to the batch of obstacles being flushed after the GL_QUADS commands for
the health bar are submitted.
This bug doesn't currently show up easily due to the obstacle batches
being small because only one texture can be bound at a time, and our
obstacles on screen rarely fit in a single image atlas. However,
upcoming changes using up to 4 textures at a time to build bigger
batches expose the problem more easily.