GL: do not reset active texture unit to 0
The active texture will be left to whatever unit was last used by struct
image. This should not matter because the new convention when running
with NO_SHADER is to disable texturing. This removes a few redundant
glActiveTexture() calls, making the GL command stream prettier, and
ensures breakage if someone breaks the new convention - which is a
feature.