translate_map_point_to_screen_pixel_func: remove NULL checks
These checks are not relied upon by any caller, and they do have a
performance impact due to this function being called so often.

The function needs to be reworked entirely (maybe made inline). Future
experiments will attempt to move it to the GPU as part of the vertex
shader.