Hide node 2 and 3 too when asking Butch for his name.
This is following what was spoken on the Usability Study, done in 2004.
The links can be found at "http://www.freedroid.org/dev-section/todo/usability-study/"

According to the study, player should not be heavily penalized for doing mistakes.
This patch omitts nodes 2 and 3 which could lead to auto-game over on later stages.
(Which is something bad!)