GL: transfer textures using a pixel buffer object (PBO)
This allows asynchronous texture transfers. Requires either OpenGL 2.1
or ARB_pixel_buffer_object, but if these are absent, the game falls back
to the previous behavior silently.

It uses one single PBO, whose content being changed when a texture is
bound. This is a first step, to be later improved by the use of
several PBOs.