GL: use non power of two textures when possible, stop padding images
We require OpenGL 2.0 due to using shaders. Most GPUs of that generation
support non power of two textures (ARB_non_power_of_two). When that is
the case, just create textures without bothering with padding.
For GPUs that still require power of two textures, modify the code so
that instead of creating a padded image and transferring that larger
image to the GPU, a texture is created with power of two dimensions but
no data - glTexImage2D with NULL argument does that since OpenGL 1.1 -
and glTexSubImage2D is use to transfer data to the part of the texture
that we care about.