Load/Delete Game menu: Fix a bug when there are more than 7 savegames.
In that case, the latest savegame is not listed.
Moreover, if there are exactly 7 savegames, 6 are listed and the 'down' option
is displayed, as expected, but nothing happens when the user presses the
'down' option.

Rewrite a bit the code handling the menu, using variables to explicitly
store the index of the special options (up, down, back).
Using them makes the code be easier to read and to fix.