[lvleditor] Fix a memleak on waypoint cut operation.
Selection - copy/cut/paste - undo/redo operations use memcpy calls to
duplicate an object when putting it in the selection_list and the clipboard_list.

There was no issue, as long as those structs only contained statically allocated
memory.

But, in some kind of objects, we replaced static arrays with dynarrays, and so
an additional operation is needed: the dynarrays contained in an object have
to be 'duplicated' when the object is memcopiedi (and have to be released when the
object is destroyed).

This patch fix the code used to handle waypoints.
The code for other objects should be reviewed...